requirement.txt
streamlit==1.26.0
audio_recorder_streamlit==0.1.0
streamlit_float==0.1.3
PyPDF2==3.0.1
langchain==0.0.258
langchain-google-genai==0.0.1
faiss-cpu==1.7.3
python-dotenv==1.0.0
google-generativeai==0.1.0
deepgram==1.0.0
groq==1.0.0